Issue-18: Added support for debugger
