+++ package.json
@@ -1,7 +1,7 @@
 {
     "name": "grunt-cucumberjs",
     "description": "Generates documentation from Cucumber features",
-    "version": "0.5.2",
+    "version": "0.6.0",
     "homepage": "https://github.com/mavdi/grunt-cucumberjs",
     "author": {
         "name": "Mehdi Avdi",
@@ -35,6 +35,7 @@
         "test": "grunt"
     },
     "devDependencies": {
+        "cucumber": "^0.4.8",
         "grunt": "~0.4.5",
         "grunt-cli": "~0.1.13",
         "grunt-contrib-clean": "~0.5.0",
@@ -48,7 +49,6 @@
         "gruntplugin"
     ],
     "dependencies": {
-        "cucumber": "~0.3.1",
         "handlebars": "~1.0.12",
         "underscore": "~1.5.2",
         "commondir": "~0.0.1"
